agent/requirements.txt
openai>=1.0.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3
